Move secondary_core to retroarch.c
