+++ griffin/griffin.c
@@ -1363,7 +1363,6 @@ MENU
 
 #ifdef HAVE_RUNAHEAD
 #include "../runahead/mem_util.c"
-#include "../runahead/secondary_core.c"
 #include "../runahead/copy_load_info.c"
 #include "../runahead/mylist.c"
 #endif
